Create base class for the genetic algorithm
